+++ src/app/delete/deleteInfo/_component/deleteInfo.module.css
