+++ src/app/delete/deleteLast/page.tsx
@@ -5,9 +5,9 @@ import { useRouter } from "next/navigation";
 
 import { useUserDataContext } from "@/context/AuthContext";
 
-import balance_together from "../../../public/balance-together.svg";
+import balance_together from "../../../../public/balance-together.svg";
 import DeleteLast from "./_component/DeleteLast";
-import DeleteLastNav from "./_component/DeleteLastNav";
+import DeleteLastNav from "../_component/DeleteLastNav";
 import styles from "./deleteLast.module.css";
 
 export default function DeleteLastPage() {
